Moved the feedback endpoint to company.py
