+++ src/companies/schemas/company.py
@@ -144,14 +144,3 @@ class UserCompanyUpdateSchema(BaseModel):
         if self.surname and not self.surname.isalpha():
             raise ValueError(TEST_ERROR_INVALID_CHARACTERS_SURNAME)
         return self
-
-
-class CompanyFeedbackCreateShema(BaseModel):
-    """Схема для создания пользователем компании обратной связи."""
-
-    question: str = Field(..., title='Задать вопрос для обратной связи')
-    # TODO: Обдумать. Скорее всего надо будет реализовать ограничение на количество символов.
-    # Схема на данный момент является по большей части заглушкой.
-
-    class Config:
-        from_attributes = True
